+++ tests/integration/test_outcome.py
@@ -1648,25 +1648,26 @@ def test_global_rate_limit(
     projectconfig = mini_sentry.add_full_project_config(project_id)
     mini_sentry.add_dsn_key_to_project(project_id)
 
+    now = datetime.utcnow().timestamp()
+
     projectconfig["config"]["quotas"] = [
         {
             "id": "test_rate_limiting" + str(uuid.uuid4()),
             "scope": "global",
             "categories": ["metric_bucket"],
             "limit": metric_bucket_limit,
-            "window": 1000,
+            # Ensures we begin at the start of a slot so we don't go to next slot in the middle of the test
+            "window": int(now),
             "reasonCode": "global rate limit hit",
         }
     ]
-
-    ts = datetime.utcnow().timestamp()
 
     def send_buckets(n):
         buckets = [
             {
                 "org_id": 1,
                 "project_id": project_id,
-                "timestamp": ts,
+                "timestamp": now,
                 "name": "d:transactions/measurements.lcp@millisecond",
                 "type": "d",
                 "value": [1.0],